Reformat pyslavseq.genome into separate Genome and Interval scripts
