+++ workflow/scripts/get_windows.py
@@ -3,7 +3,8 @@
 __author__ = 'Michael Cuoco, Rohini Gadde'
 
 import pandas as pd
-from pyslavseq.genome import Interval, Genome
+from pyslavseq.genome.Genome import Genome
+from pyslavseq.genome.Interval import Interval
 import pyslavseq.genome.interval_generator as ig
 import sys, gc, traceback
